fix: get client for network, not for celo always
